Restructure features/ directory
