+++ src/main/java/com/github/mkram17/bazaarutils/config/features/FeatureConfig.java
@@ -1,8 +1,18 @@
 package com.github.mkram17.bazaarutils.config.features;
 
-import com.github.mkram17.bazaarutils.features.*;
-import com.github.mkram17.bazaarutils.features.customorder.CustomOrder;
-import com.github.mkram17.bazaarutils.features.restrictsell.InstantSellRestrictions;
+import com.github.mkram17.bazaarutils.features.chat.StashMessages;
+import com.github.mkram17.bazaarutils.features.chat.UselessBazaarNotificationRemover;
+import com.github.mkram17.bazaarutils.features.gui.buttons.Bookmark;
+import com.github.mkram17.bazaarutils.features.gui.buttons.inputhelper.FlipHelper;
+import com.github.mkram17.bazaarutils.features.gui.buttons.inputhelper.MaxBuyOrder;
+import com.github.mkram17.bazaarutils.features.gui.buttons.inputhelper.customorder.CustomOrder;
+import com.github.mkram17.bazaarutils.features.gui.buttons.BazaarOpenOrdersButton;
+import com.github.mkram17.bazaarutils.features.gui.inventory.InstaSellHighlight;
+import com.github.mkram17.bazaarutils.features.gui.inventory.OrderStatusHighlight;
+import com.github.mkram17.bazaarutils.features.gui.overlays.OrderLimitVisual;
+import com.github.mkram17.bazaarutils.features.gui.overlays.PriceCharts;
+import com.github.mkram17.bazaarutils.features.notification.OrderFilledNotificationSound;
+import com.github.mkram17.bazaarutils.features.notification.OutbidOrderHandler;
 import com.github.mkram17.bazaarutils.utils.bazaar.market.price.PricingPosition;
 import com.teamresourceful.resourcefulconfig.api.annotations.ConfigEntry;
