+++ src/main/java/com/github/mkram17/bazaarutils/config/features/gui/InventoryConfig.java
@@ -1,15 +1,12 @@
 package com.github.mkram17.bazaarutils.config.features.gui;
 
-import com.github.mkram17.bazaarutils.features.restrictsell.InstantSellRestrictions;
-import com.github.mkram17.bazaarutils.features.restrictsell.NumericRestrictBy;
-import com.github.mkram17.bazaarutils.features.restrictsell.controls.DoubleSellRestrictionControl;
-import com.github.mkram17.bazaarutils.features.restrictsell.controls.SellRestrictionControl;
-import com.github.mkram17.bazaarutils.features.restrictsell.controls.StringSellRestrictionControl;
+import com.github.mkram17.bazaarutils.features.gui.inventory.restrictsell.InstantSellRestrictions;
+import com.github.mkram17.bazaarutils.features.gui.inventory.restrictsell.NumericRestrictBy;
+import com.github.mkram17.bazaarutils.features.gui.inventory.restrictsell.controls.DoubleSellRestrictionControl;
+import com.github.mkram17.bazaarutils.features.gui.inventory.restrictsell.controls.SellRestrictionControl;
+import com.github.mkram17.bazaarutils.features.gui.inventory.restrictsell.controls.StringSellRestrictionControl;
 import com.teamresourceful.resourcefulconfig.api.annotations.*;
-import lombok.Getter;
-import lombok.Setter;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
