+++ src/main/java/com/github/mkram17/bazaarutils/features/gui/inventory/OrderStatusHighlight.java
@@ -1,4 +1,4 @@
-package com.github.mkram17.bazaarutils.features;
+package com.github.mkram17.bazaarutils.features.gui.inventory;
 
 import com.github.mkram17.bazaarutils.config.features.DeveloperConfig;
 import com.github.mkram17.bazaarutils.events.listener.BUListener;
