+++ src/main/java/com/github/mkram17/bazaarutils/features/gui/inventory/restrictsell/NumericRestrictBy.java
@@ -1,0 +1,3 @@
+package com.github.mkram17.bazaarutils.features.gui.inventory.restrictsell;
+
+public enum NumericRestrictBy { PRICE, VOLUME }
