+++ src/main/java/com/github/mkram17/bazaarutils/features/gui/inventory/restrictsell/controls/DoubleSellRestrictionControl.java
@@ -1,6 +1,6 @@
-package com.github.mkram17.bazaarutils.features.restrictsell.controls;
+package com.github.mkram17.bazaarutils.features.gui.inventory.restrictsell.controls;
 
-import com.github.mkram17.bazaarutils.features.restrictsell.NumericRestrictBy;
+import com.github.mkram17.bazaarutils.features.gui.inventory.restrictsell.NumericRestrictBy;
 import com.github.mkram17.bazaarutils.utils.bazaar.market.order.OrderInfo;
 import com.teamresourceful.resourcefulconfig.api.annotations.Comment;
 import com.teamresourceful.resourcefulconfig.api.annotations.ConfigEntry;
