+++ src/main/java/com/github/mkram17/bazaarutils/features/gui/inventory/restrictsell/controls/SellRestrictionControl.java
@@ -1,4 +1,4 @@
-package com.github.mkram17.bazaarutils.features.restrictsell.controls;
+package com.github.mkram17.bazaarutils.features.gui.inventory.restrictsell.controls;
 
 public sealed interface SellRestrictionControl<T extends Enum<T>> extends SellRestrictor permits DoubleSellRestrictionControl, StringSellRestrictionControl {
     boolean isEnabled();
