+++ src/main/java/com/github/mkram17/bazaarutils/features/gui/inventory/restrictsell/controls/SellRestrictor.java
@@ -1,4 +1,4 @@
-package com.github.mkram17.bazaarutils.features.restrictsell.controls;
+package com.github.mkram17.bazaarutils.features.gui.inventory.restrictsell.controls;
 
 import com.github.mkram17.bazaarutils.utils.bazaar.market.order.OrderInfo;
