--- src/main/java/com/github/mkram17/bazaarutils/features/restrictsell/NumericRestrictBy.java
@@ -1,3 +1,0 @@
-package com.github.mkram17.bazaarutils.features.restrictsell;
-
-public enum NumericRestrictBy { PRICE, VOLUME }
