--- src/main/java/com/github/mkram17/bazaarutils/features/restrictsell/StringRestrictBy.java
@@ -1,3 +1,0 @@
-package com.github.mkram17.bazaarutils.features.restrictsell;
-
-public enum StringRestrictBy { NAME }
