+++ src/main/java/com/github/mkram17/bazaarutils/misc/SlotHighlightCache.java
@@ -3,7 +3,6 @@ package com.github.mkram17.bazaarutils.misc;
 import com.github.mkram17.bazaarutils.BazaarUtils;
 import com.github.mkram17.bazaarutils.config.BUConfig;
 import com.github.mkram17.bazaarutils.events.ChestLoadedEvent;
-import com.github.mkram17.bazaarutils.features.OrderStatusHighlight;
 import com.github.mkram17.bazaarutils.misc.autoregistration.RunOnInit;
 import com.github.mkram17.bazaarutils.utils.ScreenInfo;
 import meteordevelopment.orbit.EventHandler;
