+++ src/main/java/com/github/mkram17/bazaarutils/ui/SellRestrictionsMenu.java
@@ -3,10 +3,10 @@ package com.github.mkram17.bazaarutils.ui;
 import com.github.mkram17.bazaarutils.config.BUConfig;
 import com.github.mkram17.bazaarutils.config.features.gui.InventoryConfig;
 import com.github.mkram17.bazaarutils.config.util.ConfigUtil;
-import com.github.mkram17.bazaarutils.features.restrictsell.NumericRestrictBy;
-import com.github.mkram17.bazaarutils.features.restrictsell.StringRestrictBy;
-import com.github.mkram17.bazaarutils.features.restrictsell.controls.SellRestrictionControl;
-import com.github.mkram17.bazaarutils.features.restrictsell.controls.StringSellRestrictionControl;
+import com.github.mkram17.bazaarutils.features.gui.inventory.restrictsell.NumericRestrictBy;
+import com.github.mkram17.bazaarutils.features.gui.inventory.restrictsell.StringRestrictBy;
+import com.github.mkram17.bazaarutils.features.gui.inventory.restrictsell.controls.SellRestrictionControl;
+import com.github.mkram17.bazaarutils.features.gui.inventory.restrictsell.controls.StringSellRestrictionControl;
 import com.github.mkram17.bazaarutils.utils.PlayerActionUtil;
 import io.wispforest.owo.ui.base.BaseOwoScreen;
 import io.wispforest.owo.ui.component.Components;
